+++ src/shared/components/NavigationBar/index.tsx
@@ -4,7 +4,7 @@ import Link from "next/link";
 import Logo from "@/shared/components/NavigationBar/logo.svg";
 import { SearchForm } from "@/shared/components/NavigationBar/SearchForm";
 import { useLoginStore } from "@/features/auth/model/useLoginStore";
-import { Logout } from "@/features/auth/components/Logout";
+import { Logout } from "@/widgets/auth/Logout";
 import * as s from "./style.css";
 
 export const NavigationBar = () => {
